+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add ANT+ messages log file. Useful for debugging ANT+ behaviour.
 - Add configuration option for aspect ratio of videos. Fixes #64.
+- Always show distance markers in profile. Fixes #65.
 
 # 1.6.1