+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.7.1
+# 1.7.0
 
 - Add ANT+ messages log file. Useful for debugging ANT+ behaviour.
 - Add configuration option for aspect ratio of videos. Fixes #64.